--- aplication/src/App.js
@@ -1,15 +1,0 @@
-import './App.css';
-import Achado from './componentes/Achado';
-//import Login from './componentes/Login';
-//import Home from "./componentes/Home";
-//import Registo from './componentes/Registo';
-
-function App() {
-  return (
-    <div className='body'>
-      <Achado></Achado>
-    </div>
-  );
-}
-
-export default App;
