+++ aplication/src/componentes/Achado.jsx
@@ -1,13 +1,13 @@
 
 import './Achado.css'
-import Header from './Header';
-import Footer from './Footer';
+//import Header from './Header';
+//import Footer from './Footer';
 import FormAchado from './FormAchado';
 
 function Achado(){
     return(
         <div>
-                <Header></Header>
+            
                 
                 <div className="img-back"> 
         
@@ -18,7 +18,7 @@ function Achado(){
 
                  <FormAchado></FormAchado>
                 <div className='espaco'></div>
-                <Footer></Footer> 
+                
         </div>
     )
 }
