+++ aplication/src/componentes/Header.jsx
@@ -1,4 +1,5 @@
 import './Header.css';
+import { Link } from 'react-router-dom';
 function Header (){
     return (
     <>
@@ -7,16 +8,16 @@ function Header (){
 
 <img src={require("../images/Encontralogo.png")} alt="" className='image'></img>
 
-<div className='home-style'><p> Home</p></div>
+    <div className='home-style'><Link to={'/'}><p> Home</p></Link></div>
         </div>
         <div className='others'>
-        <p>Perdi</p>
-<p>Encontrei</p>
+        <a href='a' className='lin'>Perdi</a>
+        <Link to={'/Achado'}> <a href='#a' className='lin'> Encontrei</a></Link>
         
 <div className='perfil'> 
 <div className='submenu-perfil'>
     <a href='a' className="submenu-btn">Perfil</a>
-    <a href='a' className="submenu-btn">Iniciar sessao</a>
+   <Link to={'/Login'}>   <a href='a' className="submenu-btn">Iniciar sessao</a></Link>
     <a href='a' className="submenu-btn">Sair</a>
 </div>
  </div>
