+++ aplication/src/componentes/Home.jsx
@@ -1,13 +1,22 @@
 
 import "./Home.css";
-import Header from "./Header"
+
+
+
+
+
+//import Header from "./Header"
 import Slides from "./Slides"
-import Footer from "./Footer";
+//import Footer from "./Footer";
+
 function Home(){
     return (
         <> 
         <div className="content-style">
-        <Header></Header>
+            
+      
+
+
 
         <div className="img-back"> 
         
@@ -17,7 +26,6 @@ function Home(){
 
         
             <div className="home-text">Encontra aqui o que voce perdeu </div>
-            <Footer></Footer>
         </div>
 
         </>
