+++ aplication/src/componentes/Login.jsx
@@ -7,15 +7,15 @@ function Login (){
             <img id='logo' src={require("../images/Encontralogo.png")} alt="" />
             <div className='formL'>
                 <form action="">    
-                    <input type="text" placeholder='Email Institucional'/>
+                    <input className='campoText' type="text" placeholder='Email Institucional'/>
+                    
+                    <input className='campoText' type="password" placeholder='Palavra-Passe'/>
+                    
+                    <button className='bbtn'> Entrar</button>
+                    
+                    <button className='bbtn'>Registar</button>
                     <br />
-                    <input type="password" placeholder='Palavra-Passe'/>
-                    <br />
-                    <button>Entrar</button>
-                    <br />
-                    <button>Registar</button>
-                    <br /> <br />
-                    <a href=' ' >Esqueceu a sua senha?</a>
+                    <a className='fsenha' href=' ' >Esqueceu a sua senha?</a>
                 </form>
             </div>
         </div>
